# Conflicts:
#	Args/Args-Engine/Header/ECS/System.h

+++ Args/Args-Engine/Header/Time/Time.h
@@ -11,8 +11,8 @@ namespace Args
 	public:
 		Time(std::chrono::duration<double> time) : time(time){}
 
-		float Minutes() { return std::chrono::duration_cast<std::chrono::duration<float, std::ratio<60>>>(time).count(); }
-		float Seconds() { return std::chrono::duration_cast<std::chrono::duration<float>>(time).count(); }
-		float Milliseconds() { return std::chrono::duration_cast<std::chrono::duration<float, std::milli>>(time).count(); }
+		double Minutes() { return std::chrono::duration_cast<std::chrono::duration<double, std::ratio<60>>>(time).count(); }
+		double Seconds() { return std::chrono::duration_cast<std::chrono::duration<double>>(time).count(); }
+		double Milliseconds() { return std::chrono::duration_cast<std::chrono::duration<double, std::milli>>(time).count(); }
 	};
 }
