+++ Args/Args-Engine/Source/ECS/Managers/SystemManager.cpp
@@ -38,7 +38,7 @@ void Args::SystemManager::InitialiseSystems()
 
 void Args::SystemManager::UpdateSystems()
 {
-	float deltaTime = updateClock.Restart().Seconds();
+	double deltaTime = updateClock.Restart().Seconds();
 	for (auto priorityData : systemPriorities)
 		for (auto systemType : priorityData.second)
 			systems[systemType]->UpdateSystem(deltaTime);
